Fixing problem when generating dist tarball.
Had to modify the TNEF data file because it contained a file with spaces
in the name and this seemed to be an intractable problem for the
Makefiles. Because the data file was changed then baselines also needed
to be changed.

Furthermore, a previous hack to deal with the filename with a space in
it is no longer needed so it was removed.